+++ example/lib/firebase_messaging_handler.dart
@@ -1,5 +1,7 @@
 import 'package:firebase_messaging/firebase_messaging.dart';
+import 'package:flutter/foundation.dart' show debugPrint;
 import 'package:wisetrack/wisetrack.dart';
+import 'package:wisetrack_example/app_platform.dart';
 
 @pragma('vm:entry-point')
 Future<void> _firebaseMessagingBackgroundHandler(RemoteMessage message) async {
@@ -23,6 +25,10 @@ class FirebaseMessagingHandler {
     );
 
     FirebaseMessaging.onMessage.listen((RemoteMessage message) async {
+      debugPrint(
+        'new notification, title: ${message.notification?.title}, body: ${message.notification?.body}',
+      );
+      debugPrint('onMessage: ${message.data}');
       if (await WiseTrack.instance.isWiseTrackNotification(message.data)) {
         return;
       }
@@ -34,7 +40,12 @@ class FirebaseMessagingHandler {
   }
 
   static _getToken() async {
-    final token = await FirebaseMessaging.instance.getToken();
+    final token = await FirebaseMessaging.instance.getToken(
+      vapidKey:
+          AppPlatform.isWeb
+              ? '<Your VAPID Key here>'
+              : null,
+    );
     if (token != null) {
       WiseTrack.instance.setFCMToken(token);
     }