+++ example/web/firebase-messaging-sw.js
@@ -1,0 +1,32 @@
+// Import Firebase scripts
+importScripts(
+  "https://www.gstatic.com/firebasejs/10.7.1/firebase-app-compat.js"
+);
+importScripts(
+  "https://www.gstatic.com/firebasejs/10.7.1/firebase-messaging-compat.js"
+);
+importScripts("./firebase-config.js");
+
+// Initialize Firebase in the service worker
+firebase.initializeApp(firebaseConfig);
+
+// Retrieve an instance of Firebase Messaging so that it can handle background messages
+const messaging = firebase.messaging();
+
+// Handle background messages
+messaging.onBackgroundMessage(function (payload) {
+  console.log(
+    "[firebase-messaging-sw.js] Received background message ",
+    payload
+  );
+
+  const notificationTitle =
+    payload.notification?.title || "Background Message Title";
+  const notificationOptions = {
+    body: payload.notification?.body || "Background Message body.",
+    icon: "/icons/Icon-192.png",
+    badge: "/icons/Icon-192.png",
+  };
+
+  self.registration.showNotification(notificationTitle, notificationOptions);
+});