+++ lib/src/config/config.dart
@@ -1,2 +1,0 @@
-export 'channel_names.dart';
-export 'wisetrack_sdk.dart';